@@ -1,11 +1,11 @@
 name: expansion_tile_card
 description: An "expansion" on the Flutter SDK's standard ExpansionTile, to create a
   Google Material Theme inspired raised widget, ExpansionTileCard, instead.
-version: 2.2.0
+version: 3.0.0
 homepage: https://github.com/Skylled/expansion_tile_card
 
 environment:
-  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
   flutter: ">=1.26.0-0"
 
 dependencies:
