@@ -59,8 +59,7 @@ class ExpansionTileCard extends StatefulWidget {
     this.isThreeLine = false,
     this.shadowColor = const Color(0xffaaaaaa),
     this.animateTrailing = false,
-  })  : assert(initiallyExpanded != null),
-        super(key: key);
+  });
 
   final RoundedRectangleBorder? shape;
   final VisualDensity? visualDensity;
@@ -152,7 +151,7 @@ class ExpansionTileCard extends StatefulWidget {
 
   ///The color of the text of the expended card
   ///
-  ///If null, defaults to Theme.of(context).accentColor.
+  ///If null, defaults to Theme.of(context).colorScheme.secondary.
   final Color? expandedTextColor;
 
   /// The duration of the expand and collapse animations.
@@ -238,7 +237,7 @@ class ExpansionTileCardState extends State<ExpansionTileCard>
         Tween<double>(begin: widget.initialElevation, end: widget.elevation)
             .chain(_elevationTween));
     _padding = _controller.drive(_edgeInsetsTween.chain(_paddingTween));
-    _isExpanded = PageStorage.of(context)?.readState(context) as bool? ??
+    _isExpanded = PageStorage.of(context).readState(context) as bool? ??
         widget.initiallyExpanded;
     if (_isExpanded) _controller.value = 1.0;
   }
@@ -264,7 +263,7 @@ class ExpansionTileCardState extends State<ExpansionTileCard>
             });
           });
         }
-        PageStorage.of(context)?.writeState(context, _isExpanded);
+        PageStorage.of(context).writeState(context, _isExpanded);
       });
       if (widget.onExpansionChanged != null)
         widget.onExpansionChanged!(_isExpanded);
@@ -340,11 +339,11 @@ class ExpansionTileCardState extends State<ExpansionTileCard>
   void didChangeDependencies() {
     final ThemeData theme = Theme.of(context);
     _headerColorTween
-      ..begin = theme.textTheme.subtitle1!.color
-      ..end = widget.expandedTextColor ?? theme.accentColor;
+      ..begin = theme.textTheme.titleMedium!.color
+      ..end = widget.expandedTextColor ?? theme.colorScheme.secondary;
     _iconColorTween
       ..begin = theme.unselectedWidgetColor
-      ..end = widget.expandedTextColor ?? theme.accentColor;
+      ..end = widget.expandedTextColor ?? theme.colorScheme.secondary;
     _materialColorTween
       ..begin = widget.baseColor ?? theme.canvasColor
       ..end = widget.expandedColor ?? theme.cardColor;
